feat: Refactor transportation module components and update imports
