+++ src/app/components/badge/badgevertical.tsx
