+++ src/app/modules/home/components/Hero/Hero.tsx
@@ -7,11 +7,11 @@ import { DotLottieReact } from '@lottiefiles/dotlottie-react';
 import { motion, AnimatePresence, Variants } from 'framer-motion';
 
 import '@/styles/hero.css';
-import CardHero from '../../ui/cardhero/cardhero';
-import { HeroButton } from "../../ui/buttons/herobutton";
-import Badge from "../../ui/badge/badgevertical";
+import CardHero from '../cardhero/cardhero';
+import { HeroButton } from "../../../../components/buttons/herobutton";
+import Badge from "../../../../components/badge/badgevertical";
 
-import { slides } from "../../../data/slides"; // Importa los datos de los slides
+import { slides } from "../../../../data/slides"; // Importa los datos de los slides
 
 // Animations
 const container: Variants = {
